@@ -43,7 +43,7 @@ endif()
 project(STP)
 
 set(CMAKE_C_STANDARD 99)
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 
 include(GenerateExportHeader)
 include(GNUInstallDirs)
