@@ -104,9 +104,19 @@ public:
 
   DLL_PUBLIC ASTNode applySubstitutionMap(const ASTNode& n);
   DLL_PUBLIC ASTNode applySubstitutionMapUntilArrays(const ASTNode& n);
-  
-  DLL_PUBLIC ASTNode applySubstitutionMapAtTopLevel(const ASTNode& topLevel) __attribute__((warn_unused_result));
+  ASTNode applySubstitutionMapUntilArrays(const ASTNode& n, ASTNodeMap& cache);
 
+  
+  #ifdef _MSC_VER
+  #include <sal.h>
+  #define WARN_UNUSED_RESULT _Check_return_
+  #else
+  #define WARN_UNUSED_RESULT __attribute__((warn_unused_result))
+  #endif
+
+  DLL_PUBLIC WARN_UNUSED_RESULT ASTNode applySubstitutionMapAtTopLevel(const ASTNode& topLevel) WARN_UNUSED_RESULT;
+    
+  #undef WARN_UNUSED_RESULT
 
   /****************************************************************
    * Simplification functions                                     *
