@@ -45,169 +45,167 @@ namespace stp
 {
 
 using std::list;
-using std::set;
 using simplifier::constantBitP::MultiplicationStats;
 
 class Simplifier;
 class ASTNode;
-typedef vector<ASTNode> ASTVec;
 
-template <class BBNode, class BBNodeManagerT> class BitBlaster;
+using ASTVec = vector<ASTNode>;
 
-template <class BBNode, class BBNodeManagerT> class BitBlaster // not copyable
+template <class BBNode, class BBNodeManagerT> 
+class BitBlaster 
 {
+  using BBNodeVec = vector<BBNode>;
+  using BBNodeSet = std::unordered_set<BBNode>;
+
   BBNode BBTrue, BBFalse;
 
   // Memo table for bit blasted terms.  If a node has already been
   // bitblasted, it is mapped to a vector of Boolean formulas for
   // the
-  std::map<ASTNode, vector<BBNode>> BBTermMemo;
+  std::unordered_map<ASTNode, BBNodeVec, ASTNode::ASTNodeHasher, ASTNode::ASTNodeEqual> BBTermMemo;
 
   // Memo table for bit blasted formulas.  If a node has already
   // been bitblasted, it is mapped to a node representing the
   // bitblasted equivalent
-  std::map<ASTNode, BBNode> BBFormMemo;
+  std::unordered_map<ASTNode, BBNode, ASTNode::ASTNodeHasher, ASTNode::ASTNodeEqual> BBFormMemo;
 
   // Get vector of Boolean formulas for sum of two
   // vectors of Boolean formulas
-  void BBPlus2(vector<BBNode>& sum, const vector<BBNode>& y, BBNode cin);
+  void BBPlus2(BBNodeVec& sum, const BBNodeVec& y, BBNode cin);
 
   // Increment
-  vector<BBNode> BBInc(const vector<BBNode>& x);
+  BBNodeVec BBInc(const BBNodeVec& x);
 
   // Add one bit to a vector of bits.
-  vector<BBNode> BBAddOneBit(const vector<BBNode>& x, BBNode cin);
+  BBNodeVec BBAddOneBit(const BBNodeVec& x, BBNode cin);
 
   // Bitwise complement
-  vector<BBNode> BBNeg(const vector<BBNode>& x);
+  BBNodeVec BBNeg(const BBNodeVec& x);
 
   // Unary minus
-  vector<BBNode> BBUminus(const vector<BBNode>& x);
+  BBNodeVec BBUminus(const BBNodeVec& x);
 
   // Multiply.
-  vector<BBNode> BBMult(const vector<BBNode>& x, const vector<BBNode>& y,
-                        set<BBNode>& support, const ASTNode& n);
-  void mult_allPairs(const vector<BBNode>& x, const vector<BBNode>& y,
-                     set<BBNode>& support, vector<list<BBNode>>& products);
-  void mult_Booth(const vector<BBNode>& x_i, const vector<BBNode>& y_i,
-                  set<BBNode>& support, const stp::ASTNode& xN,
+  BBNodeVec BBMult(const BBNodeVec& x, const BBNodeVec& y,
+                        BBNodeSet& support, const ASTNode& n);
+  void mult_allPairs(const BBNodeVec& x, const BBNodeVec& y,
+                     BBNodeSet& support, vector<list<BBNode>>& products);
+  void mult_Booth(const BBNodeVec& x_i, const BBNodeVec& y_i,
+                  BBNodeSet& support, const stp::ASTNode& xN,
                   const stp::ASTNode& yN, vector<list<BBNode>>& products,
                   const ASTNode& n);
-  vector<BBNode> mult_normal(const vector<BBNode>& x, const vector<BBNode>& y,
-                             set<BBNode>& support, const ASTNode& n);
+  BBNodeVec mult_normal(const BBNodeVec& x, const BBNodeVec& y,
+                             BBNodeSet& support, const ASTNode& n);
 
-  vector<BBNode> batcher(const vector<BBNode>& in);
-  vector<BBNode> mergeSorted(const vector<BBNode>& in1,
-                             const vector<BBNode>& in2);
-  vector<BBNode> compareOddEven(const vector<BBNode>& in);
+  BBNodeVec batcher(const BBNodeVec& in);
+  BBNodeVec mergeSorted(const BBNodeVec& in1,
+                             const BBNodeVec& in2);
+  BBNodeVec compareOddEven(const BBNodeVec& in);
 
-  void setColumnsToZero(vector<list<BBNode>>& products, set<BBNode>& support,
+  void setColumnsToZero(vector<list<BBNode>>& products, BBNodeSet& support,
                         const ASTNode& n);
 
-  void sortingNetworkAdd(set<BBNode>& support, list<BBNode>& current,
-                         vector<BBNode>& currentSorted,
-                         vector<BBNode>& priorSorted);
+  void sortingNetworkAdd(BBNodeSet& support, list<BBNode>& current,
+                         BBNodeVec& currentSorted,
+                         BBNodeVec& priorSorted);
 
-  vector<BBNode> v6(vector<list<BBNode>>& products, set<BBNode>& support,
+  BBNodeVec v6(vector<list<BBNode>>& products, BBNodeSet& support,
                     const ASTNode& n);
-  vector<BBNode> v7(vector<list<BBNode>>& products, set<BBNode>& support,
+  BBNodeVec v7(vector<list<BBNode>>& products, BBNodeSet& support,
                     const ASTNode& n);
-  vector<BBNode> v8(vector<list<BBNode>>& products, set<BBNode>& support,
+  BBNodeVec v8(vector<list<BBNode>>& products, BBNodeSet& support,
                     const ASTNode& n);
-  vector<BBNode> v9(vector<list<BBNode>>& products, set<BBNode>& support,
+  BBNodeVec v9(vector<list<BBNode>>& products, BBNodeSet& support,
                     const ASTNode& n);
-  vector<BBNode> v13(vector<list<BBNode>>& products, set<BBNode>& support,
+  BBNodeVec v13(vector<list<BBNode>>& products, BBNodeSet& support,
                      const ASTNode& n);
 
-  vector<BBNode> multWithBounds(const ASTNode& n,
+  BBNodeVec multWithBounds(const ASTNode& n,
                                 vector<list<BBNode>>& products,
-                                set<BBNode>& toConjoinToTop);
+                                BBNodeSet& toConjoinToTop);
   bool statsFound(const ASTNode& n);
 
-  void mult_BubbleSorterWithBounds(set<BBNode>& support,
+  void mult_BubbleSorterWithBounds(BBNodeSet& support,
                                    list<BBNode>& currentColumn,
-                                   vector<BBNode>& currentSorted,
-                                   vector<BBNode>& priorSorted,
+                                   BBNodeVec& currentSorted,
+                                   BBNodeVec& priorSorted,
                                    const int minTrue = 0,
                                    const int maxTrue = ((unsigned)~0) >> 1);
 
   void buildAdditionNetworkResult(list<BBNode>& from, list<BBNode>& to,
-                                  set<BBNode>& support, const bool top,
+                                  BBNodeSet& support, const bool top,
                                   const bool empty);
-  vector<BBNode> buildAdditionNetworkResult(vector<list<BBNode>>& products,
-                                            set<BBNode>& support,
+  BBNodeVec buildAdditionNetworkResult(vector<list<BBNode>>& products,
+                                            BBNodeSet& support,
                                             const ASTNode& n);
 
-  vector<BBNode> BBAndBit(const vector<BBNode>& y, BBNode b);
+  BBNodeVec BBAndBit(const BBNodeVec& y, BBNode b);
 
   MultiplicationStats* getMS(const ASTNode& n, int& highestZero);
 
   /////////// The end of the multiplication stuff..
 
-  void checkFixed(const vector<BBNode>& v, const ASTNode& n);
+  void checkFixed(const BBNodeVec& v, const ASTNode& n);
 
   // AND each bit of vector y with single bit b and return the result.
   // (used in BBMult)
-  // vector<BBNode> BBAndBit(const vector<BBNode>& y, ASTNode b);
+  // BBNodeVec BBAndBit(const BBNodeVec& y, ASTNode b);
 
-  // Returns vector<BBNode> for result - y.  This destroys "result".
-  void BBSub(vector<BBNode>& result, const vector<BBNode>& y,
-             set<BBNode>& support);
+  // Returns BBNodeVec for result - y.  This destroys "result".
+  void BBSub(BBNodeVec& result, const BBNodeVec& y,
+             BBNodeSet& support);
 
   // build ITE's (ITE cond then[i] else[i]) for each i.
-  vector<BBNode> BBITE(const BBNode& cond, const vector<BBNode>& thn,
-                       const vector<BBNode>& els);
+  BBNodeVec BBITE(const BBNode& cond, const BBNodeVec& thn,
+                       const BBNodeVec& els);
 
   // Build a vector of zeros.
-  vector<BBNode> BBfill(unsigned int width, BBNode fillval);
+  BBNodeVec BBfill(unsigned int width, BBNode fillval);
 
   // build an EQ formula
-  BBNode BBEQ(const vector<BBNode>& left, const vector<BBNode>& right);
+  BBNode BBEQ(const BBNodeVec& left, const BBNodeVec& right);
 
   // This implements a variant of binary long division.
   // q and r are "out" parameters.  rwidth puts a bound on the
   // recursion depth.   Unsigned only, for now.
 
-  void BBDivMod(const vector<BBNode>& y, const vector<BBNode>& x,
-                vector<BBNode>& q, vector<BBNode>& r, unsigned int rwidth,
-                set<BBNode>& support);
+  void BBDivMod(const BBNodeVec& y, const BBNodeVec& x,
+                BBNodeVec& q, BBNodeVec& r, unsigned int rwidth,
+                BBNodeSet& support);
 
   // Return formula for majority function of three formulas.
   BBNode Majority(const BBNode& a, const BBNode& b, const BBNode& c);
 
   // Internal bit blasting routines.
-  BBNode BBBVLE(const vector<BBNode>& x, const vector<BBNode>& y,
+  BBNode BBBVLE(const BBNodeVec& x, const BBNodeVec& y,
                 bool is_signed, bool is_bvlt = false);
-  BBNode BBBVLE_variant1(const vector<BBNode>& x, const vector<BBNode>& y,
+  BBNode BBBVLE_variant1(const BBNodeVec& x, const BBNodeVec& y,
                          bool is_signed, bool is_bvlt = false);
-  BBNode BBBVLE_variant2(const vector<BBNode>& x, const vector<BBNode>& y,
+  BBNode BBBVLE_variant2(const BBNodeVec& x, const BBNodeVec& y,
                          bool is_signed, bool is_bvlt = false);
 
   // Return bit-blasted form for BVLE, BVGE, BVGT, SBLE, etc.
-  BBNode BBcompare(const ASTNode& form, set<BBNode>& support);
+  BBNode BBcompare(const ASTNode& form, BBNodeSet& support);
 
-  void BBLShift(vector<BBNode>& x, unsigned int shift);
-  void BBRShift(vector<BBNode>& x, unsigned int shift);
-
-  // no copy, no assign.
-  BitBlaster& operator=(const BitBlaster& other);
-  BitBlaster(const BitBlaster& other);
+  void BBLShift(BBNodeVec& x, unsigned int shift);
+  void BBRShift(BBNodeVec& x, unsigned int shift);
 
   // Checks for constants.
   void commonCheck(const ASTNode& n);
   void check(const BBNode& x, const ASTNode& n);
-  void check(const vector<BBNode>& x, const ASTNode& n);
+  void check(const BBNodeVec& x, const ASTNode& n);
 
   bool update(const ASTNode& n, const int i,
               simplifier::constantBitP::FixedBits* b, BBNode& bb,
-              set<BBNode>& support);
-  void updateTerm(const ASTNode& n, vector<BBNode>& bb, set<BBNode>& support);
-  void updateForm(const ASTNode& n, BBNode& bb, set<BBNode>& support);
+              BBNodeSet& support);
+  void updateTerm(const ASTNode& n, BBNodeVec& bb, BBNodeSet& support);
+  void updateForm(const ASTNode& n, BBNode& bb, BBNodeSet& support);
 
-  const BBNode BBForm(const ASTNode& form, set<BBNode>& support);
+  const BBNode BBForm(const ASTNode& form, BBNodeSet& support);
 
-  bool isConstant(const vector<BBNode>& v);
-  ASTNode getConstant(const vector<BBNode>& v, const ASTNode& n);
+  bool isConstant(const BBNodeVec& v);
+  ASTNode getConstant(const BBNodeVec& v, const ASTNode& n);
 
   // Nodes in this set can be replaced by their constant values, without being
   // conjoined to the top..
@@ -221,14 +219,19 @@ template <class BBNode, class BBNodeManagerT> class BitBlaster // not copyable
   ASTNodeSet booth_recoded; // Nodes that have been recoded.
 
 public:
+  BitBlaster& operator=(const BitBlaster& other) = delete;
+  BitBlaster(const BitBlaster& other) = delete;
+  ~BitBlaster() { ClearAllTables(); }
+
   simplifier::constantBitP::ConstantBitPropagation* cb;
 
   // Bit blast a bitvector term.  The term must have a kind for a
   // bitvector term.  Result is a ref to a vector of formula nodes
   // representing the boolean formula.
-  const vector<BBNode> BBTerm(const ASTNode& term, set<BBNode>& support);
-  typename std::map<ASTNode, vector<BBNode>>::iterator
-  simplify_during_bb(ASTNode& term, std::set<BBNode>& support);
+  const BBNodeVec BBTerm(const ASTNode& term, BBNodeSet& support);
+  
+  typename std::unordered_map<ASTNode, BBNodeVec, ASTNode::ASTNodeHasher, ASTNode::ASTNodeEqual>::iterator
+  simplify_during_bb(ASTNode& term, BBNodeSet& support);
 
   BitBlaster(BBNodeManagerT* bnm, Simplifier* _simp, NodeFactory* astNodeF,
              UserDefinedFlags* _uf,
@@ -248,8 +251,6 @@ public:
     BBTermMemo.clear();
     BBFormMemo.clear();
   }
-
-  ~BitBlaster() { ClearAllTables(); }
 
   // Bitblast a formula
   const BBNode BBForm(const ASTNode& form);
