@@ -27,89 +27,120 @@ THE SOFTWARE.
 
 namespace stp
 {
-// FUNC: This function builds the map between LET-var names and
-// LET-expressions
-//
-// 1. if the Let-var is already defined in the LET scope, then the
-// 1. function returns an error.
-//
-// 2. if the Let-var is already declared variable in the input, then
-// 2. the function returns an error
-//
-// 3. otherwise add the <var,letExpr> pair to the _letid_expr table.
-void LETMgr::LetExprMgr(const ASTNode& var, const ASTNode& letExpr)
+
+//Creates a new binding.  
+void LetMgr::LetExprMgr(const ASTNode& var, const ASTNode& letExpr)
 {
-  string name = var.GetName();
-  MapType::iterator it;
-  if (((it = _letid_expr_map->find(name)) != _letid_expr_map->end()))
+  if (var.GetKind() != SYMBOL)
   {
-    FatalError("LetExprMgr:The LET-var v has already been defined"
-               "in this LET scope: v =",
-               var);
+    std::cerr << var;
+    FatalError("Should be a symbol.");
   }
-
-  if (_parser_symbol_table.find(var) != _parser_symbol_table.end())
-  {
-    FatalError("LetExprMgr:This var is already declared. "
-               "cannot redeclare as a letvar: v =",
-               var);
-  }
-
   LetExprMgr(var.GetName(), letExpr);
 }
 
-void LETMgr::LetExprMgr(string name, const ASTNode& letExpr)
+//Creates a new binding.  
+void LetMgr::LetExprMgr(string name, const ASTNode& letExpr)
 {
-  assert(_letid_expr_map->find(name) == _letid_expr_map->end());
-  (*_letid_expr_map)[name] = letExpr;
+  assert(stack.size() > 0);
 
-  if (stack.size() > 0) // only smtlib2 maintains the stack.
-    stack.top().push_back(name);
+  // In CVC lets are available immediately. In SMTLIB2 it's only when the list of them has all been done.
+  if (frameMode)
+  {
+    if (interim.find(name) != interim.end())
+      {
+        string msg = "Let already created:" + name;
+        FatalError(msg.c_str());
+      }
+    interim.insert(make_pair(name,letExpr));
+  }
+  else
+  {
+    if (stack.back().find(name) != stack.back().end())
+      {
+        string msg = "Let already created:" + name;
+        FatalError(msg.c_str());
+      }
+
+    stack.back().insert(make_pair(name,letExpr));
+  }
 }
 
-// this function looks up the "var to letexpr map" and returns the
+// We're ready for these bindings to participate. 
+void LetMgr::commit()
+{
+  if (interim.size() == 0)
+    return;
+
+  for (const auto& k : interim)
+    stack.back().insert(k);
+  interim.clear();
+}
+
+
+void LetMgr::push()
+{
+  commit();
+  stack.push_back(MapType());
+}
+
+void LetMgr::pop()
+{
+  if (stack.size() == 0)
+    FatalError("Popping from empty let stack");
+
+  stack.erase(stack.end() - 1);
+}
+
+// this function looks up the symbols let-binding and returns the
 // corresponding letexpr. if there is no letexpr, then it simply
 // returns the var.
-ASTNode LETMgr::ResolveID(const ASTNode& v)
+ASTNode LetMgr::ResolveID(const ASTNode& v)
 {
   if (v.GetKind() != SYMBOL)
   {
     return v;
   }
 
-  if (_parser_symbol_table.find(v) != _parser_symbol_table.end())
-  {
-    return v;
-  }
-
-  MapType::iterator it;
-  if ((it = _letid_expr_map->find(v.GetName())) != _letid_expr_map->end())
-  {
-    return it->second;
-  }
-
+  //Lets shadow other symbols, so check them first.
+  if (isLetDeclared(v.GetName()))
+    return resolveLet(v.GetName());
+  
   return v;
 }
 
-// This function simply cleans up the LetID -> LetExpr Map.
-void LETMgr::CleanupLetIDMap(void)
+ASTNode LetMgr::resolveLet(const string s)
 {
-  while (!stack.empty())
-  {
-    stack.pop();
-  }
+  assert(isLetDeclared(s));
 
-  // std::unordered_map::clear() is very expensive on big empty maps. shortcut.
-  if (_letid_expr_map->size() == 0)
-    return;
-
-  // May contain lots of buckets, so reset.
-  delete _letid_expr_map;
-  InitializeLetIDMap();
+  // Searches backwards because they could be shadowed.
+  for (auto i = stack.rbegin(); i != stack.rend(); ++i ) 
+      if ((*i).find(s) != (*i).end())
+        return (*i).find(s)->second;
+  FatalError("never here...");
 }
 
-void LETMgr::InitializeLetIDMap(void)
+bool LetMgr::isLetDeclared(const string s)
 {
-  _letid_expr_map = new std::unordered_map<string, ASTNode>();
+  for (auto frame : stack )
+    if (frame.find(s) != frame.end())
+      return true;
+
+  return false;
+
 }
+
+void LetMgr::cleanupParserSymbolTable() 
+{
+  _parser_symbol_table.clear(); 
+}
+
+// Used only by the SMT1 & CVC parsers.
+void LetMgr::CleanupLetIDMap(void)
+{
+  interim.clear();
+  stack.clear();
+  push();
+}
+
 }
