@@ -423,14 +423,14 @@ ASTNode RemoveUnconstrained::topLevel_other(const ASTNode& n,
           // All the ASTNode vars need to map to their existing MutableASTNodes.
           // So we collect all the variables
           vector<MutableASTNode*> vars;
-          std::set<MutableASTNode*> visited;
+          std::unordered_set<MutableASTNode*> visited;
           muteOther->getAllVariablesRecursively(vars, visited);
           visited.clear();
 
-          std::map<ASTNode, MutableASTNode*> create;
+          std::unordered_map<uint64_t, MutableASTNode*> create;
           for (vector<MutableASTNode*>::iterator it = vars.begin();
                it != vars.end(); it++)
-            create.insert(std::make_pair((*it)->n, *it));
+            create.insert(std::make_pair((*it)->n.GetNodeNum(), *it));
           vars.clear();
 
           ASTNode v = bm.CreateFreshVariable(0, 0, "STP_INTERNAL_comparison");
