@@ -282,7 +282,7 @@ void ExtraMain::create_options()
       po::bool_switch(&(bm->UserFlags.print_STPinput_back_dot_flag)),
       "print dotty/neato's graph format, then exit")(
       "print-counterex,p",
-      po::bool_switch(&(bm->UserFlags.print_counterexample_flag)),
+      po::bool_switch(&(bm->UserFlags.print_counterexample_flag)), 
       "print counterexample")(
       "print-counterexbin,y",
       po::bool_switch(&(bm->UserFlags.print_binary_flag)),
@@ -375,6 +375,7 @@ void ExtraMain::create_options()
       ("check-sanity,d", 
         po::bool_switch(&(bm->UserFlags.check_counterexample_flag)),
         "construct counterexample and check it");
+
 
 #undef BOOL_ARG
 #undef INT64_ARG
